requirements.txt
langchain==0.3.8
langchain_community==0.3.8
langchain_core==0.3.21
langchain_groq==0.2.1
python-dotenv==1.0.1
streamlit==1.40.1
